requirements.txt
playwright
boto3
python-dotenv
Flask